==> src/1.dataPullAndMakeDataTidy.ipynb <==
"""
File: dataPullAndMakeDataTidy.ipynb
Description: This script automates the data pull from WRDS and S&P Global using 
             Selenium and yfinance. It cleans and tidies the downloaded data, preparing 
             it for subsequent analysis. The code handles multiple data sources and 
             formats for comprehensive financial research.
Based on Nikko Zhang's project.
Usage: Execute the script in an environment configured with Selenium, yfinance, and 
       other dependencies.
"""


## Citation/Disclaimer
Code was built off of Nikko Zhangs project found here
https://github.com/Nikkozhang/FINM32900-Final-Project/tree/main

## Table 1 Data Pull and Clean

import pandas as pd
import load_WRDS
import config
import numpy as np
from pathlib import Path
import os
import time
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import yfinance as yf

load_WRDS.main()








DATA_DIR = config.DATA_DIR
OUTPUT_DIR = config.OUTPUT_DIR

'''

# Load the data from WRDS
rcfd_series_1 = load_WRDS.load_RCFD_series_1(data_dir=DATA_DIR)
rcon_series_1 = load_WRDS.load_RCON_series_1(data_dir=DATA_DIR)
rcfd_series_2 = load_WRDS.load_RCFD_series_2(data_dir=DATA_DIR)
rcon_series_2 = load_WRDS.load_RCON_series_2(data_dir=DATA_DIR)

'''

##################################################################
###DATA PULL

###################################################################


try:
    project_dir = os.path.dirname(os.path.abspath(__file__))
except NameError:
    project_dir = os.getcwd()
download_dir = project_dir 

chrome_options = webdriver.ChromeOptions()
prefs = {"download.default_directory": download_dir}
chrome_options.add_experimental_option("prefs", prefs)


driver = webdriver.Chrome(options=chrome_options)

driver.get("https://www.spglobal.com/spdji/en/indices/fixed-income/sp-us-treasury-bond-index/")

wait = WebDriverWait(driver, 30)

try:
    cookie_button = wait.until(EC.element_to_be_clickable((By.ID, "onetrust-accept-btn-handler")))
    cookie_button.click()
    print("Cookie consent accepted.")
except Exception as e:
    print("Cookie consent not found or already dismissed:", e)

time.sleep(2)  

ten_year_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[1]/ul/li[7]/a'
ten_year_button = wait.until(EC.element_to_be_clickable((By.XPATH, ten_year_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", ten_year_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", ten_year_button)
print("Clicked the '10 Year' button.")

export_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/a'
export_button = wait.until(EC.element_to_be_clickable((By.XPATH, export_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", export_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", export_button)
print("Clicked the 'Export' button.")

downloaded_file = os.path.join(download_dir, "data.xlsx")

print("sleep time")
time.sleep(5)

driver.quit()


#############################

try:
    project_dir = os.path.dirname(os.path.abspath(__file__))
except NameError:
    project_dir = os.getcwd()
download_dir = project_dir 

chrome_options = webdriver.ChromeOptions()
prefs = {"download.default_directory": download_dir}
chrome_options.add_experimental_option("prefs", prefs)

driver = webdriver.Chrome(options=chrome_options)

driver.get("https://www.spglobal.com/spdji/en/indices/fixed-income/sp-us-treasury-bond-1-3-year-index/#overview")

wait = WebDriverWait(driver, 15)

try:
    cookie_button = wait.until(EC.element_to_be_clickable((By.ID, "onetrust-accept-btn-handler")))
    cookie_button.click()
    print("Cookie consent accepted.")
except Exception as e:
    print("Cookie consent not found or already dismissed:", e)

time.sleep(2)  
ten_year_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[1]/ul/li[7]/a'
ten_year_button = wait.until(EC.element_to_be_clickable((By.XPATH, ten_year_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", ten_year_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", ten_year_button)
print("Clicked the '10 Year' button.")

export_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/a'
export_button = wait.until(EC.element_to_be_clickable((By.XPATH, export_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", export_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", export_button)
print("Clicked the 'Export' button.")

print("1-3 downloaded")
time.sleep(5)
driver.quit()

#########################

try:
    project_dir = os.path.dirname(os.path.abspath(__file__))
except NameError:
    project_dir = os.getcwd()
download_dir = project_dir 

chrome_options = webdriver.ChromeOptions()
prefs = {"download.default_directory": download_dir}
chrome_options.add_experimental_option("prefs", prefs)

driver = webdriver.Chrome(options=chrome_options)

driver.get("https://www.spglobal.com/spdji/en/indices/fixed-income/sp-us-treasury-bond-3-5-year-index/#overview")

wait = WebDriverWait(driver, 15)

try:
    cookie_button = wait.until(EC.element_to_be_clickable((By.ID, "onetrust-accept-btn-handler")))
    cookie_button.click()
    print("Cookie consent accepted.")
except Exception as e:
    print("Cookie consent not found or already dismissed:", e)

time.sleep(2)  
ten_year_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[1]/ul/li[7]/a'
ten_year_button = wait.until(EC.element_to_be_clickable((By.XPATH, ten_year_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", ten_year_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", ten_year_button)
print("Clicked the '10 Year' button.")

export_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/a'
export_button = wait.until(EC.element_to_be_clickable((By.XPATH, export_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", export_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", export_button)
print("Clicked the 'Export' button.")

print("3-5 downloaded")
time.sleep(5)
driver.quit()

########################

try:
    project_dir = os.path.dirname(os.path.abspath(__file__))
except NameError:
    project_dir = os.getcwd()
download_dir = project_dir 

chrome_options = webdriver.ChromeOptions()
prefs = {"download.default_directory": download_dir}
chrome_options.add_experimental_option("prefs", prefs)

driver = webdriver.Chrome(options=chrome_options)

driver.get("https://www.spglobal.com/spdji/en/indices/fixed-income/sp-us-treasury-bond-7-10-year-index/#overview")

wait = WebDriverWait(driver, 15)

try:
    cookie_button = wait.until(EC.element_to_be_clickable((By.ID, "onetrust-accept-btn-handler")))
    cookie_button.click()
    print("Cookie consent accepted.")
except Exception as e:
    print("Cookie consent not found or already dismissed:", e)

time.sleep(2)  
ten_year_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[1]/ul/li[7]/a'
ten_year_button = wait.until(EC.element_to_be_clickable((By.XPATH, ten_year_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", ten_year_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", ten_year_button)
print("Clicked the '10 Year' button.")

export_xpath = '//*[@id="overview-main"]/div[1]/div[3]/div[2]/div[2]/div[1]/div[2]/div[2]/div[1]/a'
export_button = wait.until(EC.element_to_be_clickable((By.XPATH, export_xpath)))
driver.execute_script("arguments[0].scrollIntoView(true);", export_button)
time.sleep(1)
driver.execute_script("arguments[0].click();", export_button)
print("Clicked the 'Export' button.")

print("7-10 downloaded")
time.sleep(5)



driver.quit()




##################################################################
###DATA CLEAN

###################################################################

#TRES INDEX CLEAN


df = pd.read_excel('PerformanceGraphExport.xls')



df = df.drop_duplicates()

df = df.dropna(how='all')


df.columns = df.columns.str.strip().str.lower().str.replace(' ', '_')


num_cols = df.select_dtypes(include='number').columns
df[num_cols] = df[num_cols].fillna(0)

obj_cols = df.select_dtypes(include='object').columns
df[obj_cols] = df[obj_cols].fillna('')

df = df.iloc[2:, :]

df.reset_index(drop=True, inplace=True)

df.rename(columns={df.columns[0]: "date", df.columns[1]: "S&P U.S. Treasury Bond Index"}, inplace=True)

df = df.iloc[:-1, :]

df['date'] = pd.to_datetime(df['date'], errors='coerce')

start_date = pd.to_datetime("2021-02-26")
end_date = pd.to_datetime("2024-02-26")

df = df[(df['date'] >= start_date) & (df['date'] <= end_date)]

df.reset_index(drop=True, inplace=True)

df_SP_Treasury_bond_index = df
df_SP_Treasury_bond_index_1 = df_SP_Treasury_bond_index.copy()







##################################################################
#COMBINED SP13 CLEAN



import pandas as pd

# Load and clean the first dataset
df13 = pd.read_excel('PerformanceGraphExport (1).xls')

df13 = df13.drop_duplicates()
df13 = df13.dropna(how='all')

df13.columns = df13.columns.str.strip().str.lower().str.replace(' ', '_')

num_cols = df13.select_dtypes(include='number').columns
df13[num_cols] = df13[num_cols].fillna(0)

obj_cols = df13.select_dtypes(include='object').columns
df13[obj_cols] = df13[obj_cols].fillna('')

df13 = df13.iloc[2:, :]

df13.reset_index(drop=True, inplace=True)

df13.rename(columns={df13.columns[0]: "date"}, inplace=True)

df13 = df13.iloc[:-1, :]

df13.rename(columns={df13.columns[0]: "date", df13.columns[1]: "sp 1-3"}, inplace=True) 

start_date = pd.to_datetime("2019-03-08")
end_date = pd.to_datetime("2024-03-01")

df13 = df13[(df13['date'] >= start_date) & (df13['date'] <= end_date)]

df13.reset_index(drop=True, inplace=True)


##################################################################################
#COMBINED SP35 CLEAN


df35 = pd.read_excel('PerformanceGraphExport (2).xls')

df35 = df35.drop_duplicates()
df35 = df35.dropna(how='all')

df35.columns = df35.columns.str.strip().str.lower().str.replace(' ', '_')

num_cols = df35.select_dtypes(include='number').columns
df35[num_cols] = df35[num_cols].fillna(0)

obj_cols = df35.select_dtypes(include='object').columns
df35[obj_cols] = df35[obj_cols].fillna('')

df35 = df35.iloc[2:, :]

df35.reset_index(drop=True, inplace=True)

df35.rename(columns={df35.columns[0]: "date"}, inplace=True)

df35 = df35.iloc[:-1, :]

df35.rename(columns={df35.columns[0]: "date", df35.columns[1]: "sp 3-5"}, inplace=True) 

start_date = pd.to_datetime("2019-03-08")
end_date = pd.to_datetime("2024-03-01")

df35 = df35[(df35['date'] >= start_date) & (df35['date'] <= end_date)]

df35.reset_index(drop=True, inplace=True)

###########################################################
#COMBINED SP710 CLEAN



df710 = pd.read_excel('PerformanceGraphExport (3).xls')

df710 = df710.drop_duplicates()
df710 = df710.dropna(how='all')

df710.columns = df710.columns.str.strip().str.lower().str.replace(' ', '_')

num_cols = df710.select_dtypes(include='number').columns
df710[num_cols] = df710[num_cols].fillna(0)

obj_cols = df710.select_dtypes(include='object').columns
df710[obj_cols] = df710[obj_cols].fillna('')

df710 = df710.iloc[2:, :]

df710.reset_index(drop=True, inplace=True)

df710.rename(columns={df710.columns[0]: "date"}, inplace=True)

df710 = df710.iloc[:-1, :]

df710.rename(columns={df710.columns[0]: "date", df710.columns[1]: "sp 7-10"}, inplace=True) 

start_date = pd.to_datetime("2019-03-08")
end_date = pd.to_datetime("2024-03-01")

df710 = df710[(df710['date'] >= start_date) & (df710['date'] <= end_date)]

df710.reset_index(drop=True, inplace=True)


print(df13)
print(df35)
print(df710)


###############################################################################################
#YFINANCE iShares
tickers = {
    "iShares 0-1": "IB01.L",   # iShares Short Treasury Bond ETF
    "iShares 1-3": "SHY",   # iShares 1-3 Year Treasury Bond ETF
    "iShares 7-10": "IEF",  # iShares 7-10 Year Treasury Bond ETF
    "iShares 10-20": "TLH", # iShares 10-20 Year Treasury Bond ETF
    "iShares 20+": "TLT"    # iShares 20+ Year Treasury Bond ETF
}

# Define date range
start_date = "2019-03-08"
end_date = "2024-03-01"

# Fetch data
iShares = yf.download(list(tickers.values()), start=start_date, end=end_date, progress=False)["Close"]

# Rename columns to match requested format
iShares.columns = list(tickers.keys())

# Reset index and rename date column
iShares.reset_index(inplace=True)
iShares.rename(columns={"Date": "date"}, inplace=True)

iShares['iShares 1-3'] = iShares['iShares 1-3'] + 3.007607



iShares['iShares 7-10'] = iShares['iShares 7-10'] + 3.420052


iShares['iShares 10-20'] = iShares['iShares 10-20'] + 5.018005


iShares['iShares 20+'] = iShares['iShares 20+'] + 4.322273




################################################################################################
#JOINING ISHARES AND COMBINED PT 1,2,3


# Convert the 'date' column in all DataFrames to datetime format
dfs = [iShares, df13, df35, df710]
for df in dfs:
    df['date'] = pd.to_datetime(df['date'])

# Merge all DataFrames on the 'date' column using an outer join
treasury_prices = iShares.merge(df13, on='date', how='outer') \
                   .merge(df35, on='date', how='outer') \
                   .merge(df710, on='date', how='outer')


treasury_prices = treasury_prices.set_index("date").asfreq("D").ffill()
treasury_prices = treasury_prices.reset_index()

treasury_prices_1 = treasury_prices.copy()




#treasury_prices = pd.read_excel("./data/manual/combined_index_df.xlsx")


#treasury_prices_1 = treasury_prices.copy()


###################################################################################
#YFINANCE MBS automation

import yfinance as yf
import pandas as pd

# Define the ticker symbol
ticker_symbol = "MBB"

# Set the date range based on your dataset
start_date = "2019-03-11"
end_date = "2024-03-08"

# Fetch historical market data for MBB within the same date range
df_iShare_MBS_ETF = yf.download(ticker_symbol, start=start_date, end=end_date)
df_iShare_MBS_ETF.columns = df_iShare_MBS_ETF.columns.get_level_values(0)
df_iShare_MBS_ETF.rename(columns={"Close": "Adj Close"}, inplace=True)
df_iShare_MBS_ETF = df_iShare_MBS_ETF.reset_index()
df_iShare_MBS_ETF_1 = df_iShare_MBS_ETF.copy()

####################################################################################

### Save data

treasury_prices.to_csv("Tres.csv", index=False)
df_iShare_MBS_ETF.to_csv("MBS.csv", index=False)
df_SP_Treasury_bond_index.to_csv("TresIndex.csv",index=False)

## Table 1A/Figure 1A Data Pull and Clean

from pathlib import Path
OUTPUT_DIR = Path(config.OUTPUT_DIR)
DATA_DIR = Path(config.DATA_DIR)
WRDS_USERNAME = config.WRDS_USERNAME

START_DATE = config.START_DATE
END_DATE = config.END_DATE
from datetime import datetime
from dateutil.relativedelta import relativedelta
from pathlib import Path
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import wrds
from scipy.stats.mstats import winsorize

db = wrds.Connection(wrds_username=WRDS_USERNAME)

rcon_1_query = f"""
    SELECT 
        rcon_1.rssd9999, rcon_1.rssd9001,
        rcon_1.RCON0071, rcon_1.RCON1773, rcon_1.RCONHT55,
        rcon_1.RCONHT57, rcon_1.RCONG309, rcon_1.RCONG311,
        rcon_1.RCONG313, rcon_1.RCONG315, rcon_1.RCONG317,
        rcon_1.RCONG319, rcon_1.RCONG321, rcon_1.RCONG323,
        rcon_1.RCONK143, rcon_1.RCONK145, rcon_1.RCONK147,
        rcon_1.RCONK149, rcon_1.RCONK151, rcon_1.RCONK153,
        rcon_1.RCONK155, rcon_1.RCONK157, rcon_1.RCONC988,
        rcon_1.RCONC027, rcon_1.RCONHT59, rcon_1.RCONHT61,
        rcon_1.RCON1743, rcon_1.RCON1746, rcon_1.RCONF158,
        rcon_1.RCONF159, rcon_1.RCON5367, rcon_1.RCON5368,
        rcon_1.RCONF160, rcon_1.RCONF161, rcon_1.RCON1590,
        rcon_1.RCON1766, rcon_1.RCONB538, rcon_1.RCONK137,
        rcon_1.RCONK207, rcon_1.rconj454, rcon_1.RCONJ454,
       
        rcon_1.RCONB987, rcon_1.RCONJ451,

        rcon_1.rconmt91, rcon_1.rconmt87, rcon_1.rconhk14, rcon_1.rconhk15,
        rcon_1.rconb993, rcon_1.rcon3230

   
    FROM 
        bank_all.wrds_call_rcon_1 AS rcon_1
    
    
    WHERE 
        rcon_1.rssd9999  = '{END_DATE}' 
    """
rcon_2_query = f"""
    SELECT 
        rcon_2.rssd9001,rcon_2.rcon0081, rcon_2.rcon1771, rcon_2.rcon0213, 
        rcon_2.rcon1287, rcon_2.rcon1738, rcon_2.rcon1741, rcon_2.rcon2122,
        rcon_2.rcon1420, rcon_2.rcon1797, rcon_2.rcon1460, rcon_2.rconb539,
        rcon_2.rconj464, rcon_2.rconb989,

        rcon_2.rcon2200, rcon_2.rconhk05, rcon_2.rconj474, rcon_2.rconb995,
        rcon_2.rconk222, rcon_2.rcon2948, rcon_2.rcon2930, rcon_2.rcong105,
        rcon_2.rcon3838, rcon_2.rcon3632, rcon_2.rcon2170


    FROM 
        bank_all.wrds_call_rcon_2 AS rcon_2
    
    
    WHERE 
        rcon_2.rssd9999  = '{END_DATE}' 
    """


rcfd_1_query = f"""
    SELECT 
        rcfd_1.rssd9001, rcfd_1.rssd9999,
        rcfd_1.rcfd0010,rcfd_1.rcfd1773,rcfd_1.rcfdg301,rcfd_1.rcfdg303,
        rcfd_1.rcfdg305,rcfd_1.rcfdg307,rcfd_1.rcfdg309,rcfd_1.rcfdg311,
        rcfd_1.rcfdg313,rcfd_1.rcfdg315,rcfd_1.rcfdg317,rcfd_1.rcfdg319,
        rcfd_1.rcfdg321,rcfd_1.rcfdg323,rcfd_1.rcfdk143,rcfd_1.rcfdk145,
        rcfd_1.rcfdk147,rcfd_1.rcfdk149, rcfd_1.rcfdk151,rcfd_1.rcfdk153,
        rcfd_1.rcfdk155,rcfd_1.rcfdk157, rcfd_1.rcfdc988,rcfd_1.rcfdc027,
        rcfd_1.RCFD1738,rcfd_1.RCFD1741, rcfd_1.RCFD1743,rcfd_1.RCFD1746,
        rcfd_1.rcfdf158,rcfd_1.rcfdf159, rcfd_1.rcfd5367,rcfd_1.rcfd5368,
        rcfd_1.rcfdf160,rcfd_1.rcfdf161, rcfd_1.rcfd1590,rcfd_1.rcfd1763,
        rcfd_1.rcfd1764,rcfd_1.rcfdb538, rcfd_1.rcfdb539,rcfd_1.rcfdk137,
        rcfd_1.rcfdk137,rcfd_1.rcfdk207,

        rcfd_1.rcfd2930,rcfd_1.rcfd3230

       

  
       

    FROM 
        bank_all.wrds_call_rcfd_1 AS rcfd_1
    
    
    WHERE 
        rcfd_1.rssd9999  = '{END_DATE}' 
    """
rcfd_2_query = f"""
    SELECT 
        rcfd_2.rssd9001, 
        rcfd_2.rcfd1771, rcfd_2.rcfd0213, rcfd_2.rcfd1287, rcfd_2.rcfd2122,
        rcfd_2.rcfd1420, rcfd_2.rcfd1797, rcfd_2.rcfd1460, rcfd_2.rcfdb989,

        rcfd_2.rcfd2948, rcfd_2.rcfdg105, rcfd_2.rcfd3838, rcfd_2.rcfd3632,
        rcfd_2.rcfd2170

    FROM 
        bank_all.wrds_call_rcfd_2 AS rcfd_2
    
    
    WHERE 
        rcfd_2.rssd9999  = '{END_DATE}' 
    """

rcfn_query = f"""
    SELECT 
        rcfn_1.rssd9001, rcfn_1.rcfn2200

    FROM 
        bank_all.wrds_call_rcfn_1 AS rcfn_1
    
    
    WHERE 
        rcfn_1.rssd9999  = '{END_DATE}' 
    """



##################################################################
###DATA CLEAN

###################################################################


rcon_1_df = db.raw_sql(rcon_1_query)
rcon_1_df = rcon_1_df.sort_values(by = 'rssd9001').set_index('rssd9001')

rcon_2_df = db.raw_sql(rcon_2_query)
rcon_2_df = rcon_2_df.sort_values(by = 'rssd9001').set_index('rssd9001')

rcon_df = pd.merge(rcon_1_df, rcon_2_df, left_index=True, right_index=True, how='inner')





rcfd_1_df = db.raw_sql(rcfd_1_query)
rcfd_1_df = rcfd_1_df.sort_values(by = 'rssd9001').set_index('rssd9001')
rcfd_2_df = db.raw_sql(rcfd_2_query)
rcfd_2_df = rcfd_2_df.sort_values(by = 'rssd9001').set_index('rssd9001')
rcfd_df = pd.merge(rcfd_1_df, rcfd_2_df, left_index=True, right_index=True, how='inner')




rcfn_df = db.raw_sql(rcfn_query)









global_rmbs = ['rcfdg301','rcfdg303','rcfdg305','rcfdg307','rcfdg309','rcfdg311',
               'rcfdg313','rcfdg315','rcfdg317','rcfdg319','rcfdg321','rcfdg323']
global_cmbs = ['rcfdk143','rcfdk145','rcfdk147','rcfdk149','rcfdk151','rcfdk153',
               'rcfdk157']
global_abs = ['rcfdc988','rcfdc027']
global_other = ['rcfd1738','rcfd1741','rcfd1743','rcfd1746']
global_rs_loan = ['rcfdf158','rcfdf159','rcfd1420','rcfd1420','rcfd1797','rcfd5367','rcfd5368','rcfd1460','rcfdf160','rcfdf161']
global_rs_residential_loan = ['rcfd1420', 'rcfd1797','rcfd5367','rcfd5368','rcfd1460']
global_rs_commerical_loan = ['rcfdf160','rcfdf161']
global_rs_other_loan = ['rcfdf158','rcfdf159']
global_ci_loan = ['rcfd1763','rcfd1764']
global_consumer_loan = ['rcfdb538','rcfdb539','rcfdk137','rcfdk207']



domestic_cash = ['rcon0081','rcon0071']
domestic_total = ['rcon1771','rcon1773']
domestic_treasury = ['rcon0213','rcon1287']
domestic_rmbs = ['rconht55','rconht57','rcong309','rcong311',
               'rcong313','rcong315','rcong317','rcong319','rcong321','rcong323']
domestic_cmbs = ['rconk143','rconk145','rconk147','rconk149','rconk151','rconk153',
               'rconk157']
domestic_abs = ['rconc988','rconc027','rconht59','rconht61']
domestic_other = ['rcon1738','rcon1741','rcon1743','rcon1746']
domestic_rs_loan = ['rconf158','rconf159','rcon1420','rcon1420','rcon1797','rcon5367','rcon5368','rcon1460','rconf160','rconf161']
domestic_rs_residential_loan = ['rcon1420','rcon1797','rcon5367','rcon5368','rcon1460']
domestic_rs_commerical_loan = ['rconf160','rconf161']
domestic_rs_other_loan = ['rconf158','rconf159']
domestic_ci_loan = ['rcon1766']
domestic_consumer_loan = ['rconb538','rconb539','rconk137','rconk207']
domestic_non_rep_loan = ['rconj454','rconj464','rconj451']



insured_deposit = ['rconhk05','rconmt91','rconmt87']
uninsured_long = ['rconhk14','rconhk15']




rcfd_data = pd.DataFrame(index= rcfd_df.index)
rcfd_data['Total Asset'] = rcfd_df['rcfd2170']
rcfd_data['cash'] = rcfd_df['rcfd0010']
rcfd_data['security_total'] = rcfd_df['rcfd1771'] + rcfd_df['rcfd1773']
rcfd_data['security_treasury'] = rcfd_df['rcfd0213'] + rcfd_df['rcfd1287']
rcfd_data['security_rmbs'] = rcfd_df[global_rmbs].sum(axis=1)
rcfd_data['security_cmbs'] = rcfd_df[global_cmbs].sum(axis=1)
rcfd_data['security_abs'] = rcfd_df[global_abs].sum(axis=1)
rcfd_data['security_other'] = rcfd_df[global_other].sum(axis=1)
rcfd_data['Total_Loan'] = rcfd_df['rcfd2122']
rcfd_data['Real_Estate_Loan'] = rcfd_df[global_rs_loan].sum(axis=1)
rcfd_data['Residential_Mortgage'] = rcfd_df[global_rs_residential_loan].sum(axis=1)
rcfd_data['Commerical_Mortgage'] = rcfd_df[global_rs_commerical_loan].sum(axis=1)
rcfd_data['Other_Real_Estate_Mortgage'] = rcfd_df[global_rs_other_loan].sum(axis=1)
rcfd_data['Agri_Loan'] = rcfd_df['rcfd1590']
rcfd_data['Comm_Indu_Loan'] = rcfd_df[global_ci_loan].sum(axis=1)
rcfd_data['Consumer_Loan'] = rcfd_df[global_consumer_loan].sum(axis=1)
rcfd_data['Non_Rep_Loan'] =  np.nan
rcfd_data['Fed_Fund_Sold'] = rcon_df['rconb987']
rcfd_data['Reverse_Repo'] = rcfd_df['rcfdb989']


rcon_data = pd.DataFrame(index= rcon_df.index)
rcon_data['Total Asset'] = rcon_df['rcon2170']
rcon_data['cash'] = rcon_df[domestic_cash].sum(axis = 1)
rcon_data['security_total'] = rcon_df[domestic_total].sum(axis = 1)
rcon_data['security_treasury'] = rcon_df[domestic_treasury].sum(axis = 1)
rcon_data['security_rmbs'] = rcon_df[domestic_rmbs].sum(axis=1)
rcon_data['security_cmbs'] = rcon_df[domestic_cmbs].sum(axis=1)
rcon_data['security_abs'] = rcon_df[domestic_abs].sum(axis=1)
rcon_data['security_other'] = rcon_df[domestic_other].sum(axis=1)
rcon_data['Total_Loan'] = rcon_df['rcon2122']
rcon_data['Real_Estate_Loan'] = rcon_df[domestic_rs_loan].sum(axis=1)
rcon_data['Residential_Mortgage'] = rcon_df[domestic_rs_residential_loan].sum(axis=1)
rcon_data['Commerical_Mortgage'] = rcon_df[domestic_rs_commerical_loan].sum(axis=1)
rcon_data['Other_Real_Estate_Mortgage'] = rcon_df[domestic_rs_other_loan].sum(axis=1)
rcon_data['Agri_Loan'] = rcon_df['rcon1590']
rcon_data['Comm_Indu_Loan'] = rcon_df[domestic_ci_loan].sum(axis=1)
rcon_data['Consumer_Loan'] = rcon_df[domestic_consumer_loan].sum(axis=1)
rcon_data['Non_Rep_Loan'] =  rcon_df[domestic_non_rep_loan].sum(axis=1)
rcon_data['Fed_Fund_Sold'] = rcon_df['rconb987']
rcon_data['Reverse_Repo'] = rcon_df['rconb989']




bank_asset = pd.merge(rcfd_data, rcon_data, left_index=True, right_index=True, how='outer', suffixes=('', '_df2'))
replace_index = bank_asset[bank_asset['cash'].isna()].index

bank_asset.loc[replace_index, bank_asset.columns[:19]] = bank_asset.loc[replace_index, bank_asset.columns[19:]].values
bank_asset.drop(columns='Non_Rep_Loan', inplace= True)
bank_asset.rename(columns = {'Non_Rep_Loan_df2':'Non_Rep_Loan'}, inplace= True)
columns_to_drop = [col for col in bank_asset.columns if '_df2' in col]
bank_asset.drop(columns=columns_to_drop, inplace= True)



global_liability = pd.DataFrame(index= rcon_df.index)
global_liability['Total Liability'] = rcfd_df['rcfd2948']
global_liability['Domestic Deposit'] = rcon_df['rcon2200']
global_liability['Insured Deposit'] = rcon_df[insured_deposit].sum(axis=1)
global_liability['Uninsured Deposit'] = global_liability['Domestic Deposit'] - global_liability['Insured Deposit']
global_liability['Uninsured Time Deposits'] = rcon_df['rconj474']
global_liability['Uninsured Long-Term Time Deposits'] = rcon_df[uninsured_long].sum(axis=1)
global_liability['Uninsured Short-Term Time Deposits'] = rcon_df['rconk222']
global_liability['Foreign Deposit'] = rcfn_df['rcfn2200']
global_liability['Fed Fund Purchase'] = rcon_df['rconb993']
global_liability['Repo'] = rcon_df['rconb995']
global_liability['Other Liability'] = rcfd_df['rcfd2930']
global_liability['Total Equity'] = rcfd_df['rcfdg105']
global_liability['Common Stock'] = rcfd_df['rcfd3230']
global_liability['Preferred Stock'] = rcfd_df['rcfd3838']
global_liability['Retained Earning'] = rcfd_df['rcfd3632']


domestic_liability = pd.DataFrame(index= rcon_df.index)
domestic_liability['Total Liability'] = rcon_df['rcon2948']
domestic_liability['Domestic Deposit'] = rcon_df['rcon2200']
domestic_liability['Insured Deposit'] = rcon_df[insured_deposit].sum(axis=1)
domestic_liability['Uninsured Deposit'] = domestic_liability['Domestic Deposit'] - domestic_liability['Insured Deposit']
domestic_liability['Uninsured Time Deposits'] = rcon_df['rconj474']
domestic_liability['Uninsured Long-Term Time Deposits'] = rcon_df[uninsured_long].sum(axis=1)
domestic_liability['Uninsured Short-Term Time Deposits'] = rcon_df['rconk222']
domestic_liability['Foreign Deposit'] = rcfn_df['rcfn2200']
domestic_liability['Fed Fund Purchase'] = rcon_df['rconb993']
domestic_liability['Repo'] = rcon_df['rconb995']
domestic_liability['Other Liability'] = rcon_df['rcon2930']
domestic_liability['Total Equity'] = rcon_df['rcong105']
domestic_liability['Common Stock'] = rcon_df['rcon3230']
domestic_liability['Preferred Stock'] = rcon_df['rcon3838']
domestic_liability['Retained Earning'] = rcon_df['rcon3632']




bank_liability = pd.merge(global_liability, domestic_liability, left_index=True, right_index=True, how='outer', suffixes=('', '_df2'))
replace_index = bank_liability[bank_liability['Total Liability'].isna()].index

bank_liability.loc[replace_index, bank_liability.columns[:15]] = bank_liability.loc[replace_index, bank_liability.columns[15:]].values
columns_to_drop = [col for col in bank_liability.columns if '_df2' in col]
bank_liability.drop(columns=columns_to_drop, inplace= True)





### Save data

bank_asset.to_csv('WRDSasset.csv')

bank_liability.to_csv('WRDSliability.csv')
