==> src/2.briefTourOfData.ipynb <==
"""
File: briefTourOfData.ipynb
Description: This script offers a brief tour of the cleaned WRDS asset and liability 
             data used to build Table A1 and Figure A1. It loads the datasets and displays 
             key metrics to provide an initial overview of the data structure and quality.
Usage: Run the script to perform preliminary data exploration.
"""


import pandas as pd

Cleaned WRDS data used to build asset part of Table A1 & Figure A1

df = pd.read_csv('WRDSasset.csv')
df

Cleaned WRDS data used to build liability part of Table A1 & Figure A1

df = pd.read_csv('WRDSliability.csv')
df

Treasury bond indexes of various maturity lengths from S&P and BlackRock (iShares) are used to calculate Table 1

df = pd.read_csv('Tres.csv')
df

Single treasury bond index is used to further calculate Table 1

df = pd.read_csv('TresIndex.csv')
df

Single mortgage backed bond index is used to further calculate Table 1

df = pd.read_csv('MBS.csv')
df
